LIAVIDEOS/app/  requirements.txt
torch>=2.2.0
whisperx>=3.1.1
openai-whisper>=20231117
ffmpeg-python>=0.2.0
moviepy>=1.0.3
sentence-transformers>=3.0.1
vadersentiment>=3.3.2
librosa>=0.10.2.post1
numpy>=1.26.4
pandas>=2.2.2
tqdm>=4.66.4
opencv-python>=4.10.0.84
spacy>=3.7.5
pydub>=0.25.1
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
